feat(recruit): working on update recruit tests
