+++ tests/sanity/tests/actions.spec.ts
@@ -38,6 +38,7 @@ test.describe('actions tests', () => {
 
     await expect(page).toHaveURL(`${PlatformURI}/workbench/sanity-ws/recruit/vacancies`)
   })
+
   test('action-switch-applications', async ({ page }) => {
     await page.click('[id="app-recruit\\:string\\:RecruitApplication"]')
